[enh] Prepare config panel for domain
